Reorganize Native API middleware/handlers
